sql/catalog/resolver,tree: move most of name resolution out of tree
Release note: None
