+++ pkg/sql/faketreeeval/evalctx.go
@@ -71,11 +71,11 @@ func (so *DummySequenceOperators) ResolveTableName(
 	return 0, errors.WithStack(errSequenceOperators)
 }
 
-// LookupSchema is part of the tree.EvalDatabase interface.
-func (so *DummySequenceOperators) LookupSchema(
+// SchemaExists is part of the tree.EvalDatabase interface.
+func (so *DummySequenceOperators) SchemaExists(
 	ctx context.Context, dbName, scName string,
-) (bool, tree.SchemaMeta, error) {
-	return false, nil, errors.WithStack(errSequenceOperators)
+) (bool, error) {
+	return false, errors.WithStack(errSequenceOperators)
 }
 
 // IsTableVisible is part of the tree.EvalDatabase interface.
@@ -236,11 +236,9 @@ func (ep *DummyEvalPlanner) ParseQualifiedTableName(sql string) (*tree.TableName
 	return parser.ParseQualifiedTableName(sql)
 }
 
-// LookupSchema is part of the tree.EvalDatabase interface.
-func (ep *DummyEvalPlanner) LookupSchema(
-	ctx context.Context, dbName, scName string,
-) (bool, tree.SchemaMeta, error) {
-	return false, nil, errors.WithStack(errEvalPlanner)
+// SchemaExists is part of the tree.EvalDatabase interface.
+func (ep *DummyEvalPlanner) SchemaExists(ctx context.Context, dbName, scName string) (bool, error) {
+	return false, errors.WithStack(errEvalPlanner)
 }
 
 // IsTableVisible is part of the tree.EvalDatabase interface.
