+++ pkg/sql/sem/builtins/builtins.go
@@ -3616,7 +3616,7 @@ may increase either contention or retry errors, or both.`,
 				curDb := evalCtx.SessionData.Database
 				iter := evalCtx.SessionData.SearchPath.IterWithoutImplicitPGSchemas()
 				for scName, ok := iter.Next(); ok; scName, ok = iter.Next() {
-					if found, _, err := evalCtx.Planner.LookupSchema(ctx, curDb, scName); found || err != nil {
+					if found, err := evalCtx.Planner.SchemaExists(ctx, curDb, scName); found || err != nil {
 						if err != nil {
 							return nil, err
 						}
@@ -3660,7 +3660,7 @@ may increase either contention or retry errors, or both.`,
 					iter = evalCtx.SessionData.SearchPath.IterWithoutImplicitPGSchemas()
 				}
 				for scName, ok := iter.Next(); ok; scName, ok = iter.Next() {
-					if found, _, err := evalCtx.Planner.LookupSchema(ctx, curDb, scName); found || err != nil {
+					if found, err := evalCtx.Planner.SchemaExists(ctx, curDb, scName); found || err != nil {
 						if err != nil {
 							return nil, err
 						}
