+++ pkg/sql/sem/tree/BUILD.bazel
@@ -212,7 +212,6 @@ go_test(
         "//pkg/security",
         "//pkg/security/securitytest",
         "//pkg/settings/cluster",
-        "//pkg/sql/catalog/resolver",
         "//pkg/sql/catalog/typedesc",
         "//pkg/sql/oidext",
         "//pkg/sql/opt/exec/execbuilder",
